+++ api/movie.php
@@ -1,46 +1,40 @@
 <?php
 
-	require_once "database_connection.php";
-	require_once "jwt.php";
+	require_once "helpers/jwt.php";
+	require_once "helpers/reviews/check_review_data.php";
+	require_once "helpers/reviews/movie_and_review_check.php";
 
-	function route($method, $urlList, $requestData) {
+	function route($method, $urlList, $requestData): void {
 		global $link;
 		switch ($method) {
 			case "POST":
+				if (count($urlList) != 5) {
+					setHTTPStatus('404', 'Missing resource is requested');
+					return;
+				}
 				$token = substr(getallheaders()['Authorization'], 7);
 				$isLogoutToken = $link->query("SELECT user_id FROM tokens WHERE value LIKE '$token'")->fetch_assoc();
-				if (!isExpired($token) && $isLogoutToken == null) {
+				if (isValid($token) && !isExpired($token) && $isLogoutToken == null) {
 					$usernameFromToken = getPayload($token)['unique_name'];
 					$user = $link->query("SELECT user_id FROM users WHERE username='$usernameFromToken'")->fetch_assoc();
 					$userID = $user['user_id'];
-					$review = $link->query("SELECT review_id FROM reviews WHERE user_id='$userID'")->fetch_assoc();
 					$movieID = $urlList[2];
+					$checkIsMovieExist = $link->query("SELECT * FROM movies WHERE movie_id='$movieID'")->fetch_assoc();
+					if (!$checkIsMovieExist) {
+						setHTTPStatus('404',"There is no movie with this '$movieID' identifier");
+						return;
+					}
+					$review = $link->query("SELECT review_id FROM reviews WHERE user_id='$userID' AND movie_id='$movieID'")->fetch_assoc();
 					if (!$review) {
-						$isValidated = true;
-						$validationErrors = [];
 						$reviewText = $requestData->body->reviewText;
 						$rating = $requestData->body->rating;
-						if ($rating < 0 || $rating > 10 || !is_int($rating)) {
-							$validationErrors[] = ["Rating" => 'The field Rating must be between 0 and 10'];
-							$isValidated = false;
-						}
 						$isAnonymous = $requestData->body->isAnonymous;
-						if (!is_bool($isAnonymous)) {
-							$validationErrors[] = ["IsAnonymous" => 'The field IsAnonymous can only be true or false'];
-							$isValidated = false;
-						}
-						if (!$isValidated) {
-							$messageResult = array(
-								'message' => 'Adding review is failed',
-								'errors' => []
-							);
-							$messageResult['errors'] = $validationErrors;
-							setHTTPStatus('400', $messageResult);
+						if (!isValidReviewData($rating, $isAnonymous, $reviewText)) {
 							return;
 						}
-						$now = DateTime::createFromFormat('U.u', number_format(microtime(true), 6, '.', ''));
-						$nowFormatted = substr($now->format('Y-m-d\TH:i:s.u'), 0, -3) . 'Z';
 						$isAnonymous = (int)$isAnonymous;
+						$now = DateTime::createFromFormat('U.u', number_format(microtime(true), 6, '.', ''));
+						$nowFormatted = str_replace(["T", "Z"], " ", trim(substr($now->format('Y-m-d\TH:i:s.u'), 0, -3) . 'Z'));
 						$reviewInsertResult = $link->query("INSERT INTO reviews(review_id, user_id, movie_id, review_text, rating, is_anonymous, create_datetime) 
 														VALUES(UUID(), '$userID', '$movieID', '$reviewText', '$rating', '$isAnonymous', '$nowFormatted')");
 					}
@@ -60,47 +54,24 @@
 				}
 				$token = substr(getallheaders()['Authorization'], 7);
 				$isLogoutToken = $link->query("SELECT user_id FROM tokens WHERE value LIKE '$token'")->fetch_assoc();
-				if (!isExpired($token) && $isLogoutToken == null) {
+				if (isValid($token) && !isExpired($token) && $isLogoutToken == null) {
 					$usernameFromToken = getPayload($token)['unique_name'];
 					$user = $link->query("SELECT user_id FROM users WHERE username='$usernameFromToken'")->fetch_assoc();
 					$userID = $user['user_id'];
 					$movieID = $urlList[2];
 					$reviewID = $urlList[4];
-					$reviewCheck = $link->query("SELECT review_id FROM reviews WHERE review_id='$reviewID' AND user_id='$userID'")->fetch_assoc();
-					$movieCheck = $link->query("SELECT review_id FROM reviews WHERE user_id='$userID' AND movie_id='$movieID'")->fetch_assoc();
-					if (!$movieCheck) {
-						setHTTPStatus('404','The user has no review on this movie');
+					if (!checkIfTheUserHasReview($movieID, $userID) || !checkIfTheUserCanChangeReview($reviewID, $userID)) {
 						return;
 					}
-					if (!$reviewCheck) {
-						setHTTPStatus('403','User cannot change foreign review');
-						return;
-					}
-					$isValidated = true;
-					$validationErrors = [];
 					$reviewText = $requestData->body->reviewText;
 					$rating = $requestData->body->rating;
-					if ($rating < 0 || $rating > 10 || !is_int($rating)) {
-						$validationErrors[] = ["Rating" => 'The field Rating must be between 0 and 10'];
-						$isValidated = false;
-					}
 					$isAnonymous = $requestData->body->isAnonymous;
-					if (!is_bool($isAnonymous)) {
-						$validationErrors[] = ["IsAnonymous" => 'The field IsAnonymous can only be true or false'];
-						$isValidated = false;
-					}
-					if (!$isValidated) {
-						$messageResult = array(
-							'message' => 'Adding review is failed',
-							'errors' => []
-						);
-						$messageResult['errors'] = $validationErrors;
-						setHTTPStatus('400', $messageResult);
+					if (!isValidReviewData($rating, $isAnonymous, $reviewText)) {
 						return;
 					}
-					$now = DateTime::createFromFormat('U.u', number_format(microtime(true), 6, '.', ''));
-					$nowFormatted = substr($now->format('Y-m-d\TH:i:s.u'), 0, -3) . 'Z';
 					$isAnonymous = (int)$isAnonymous;
+					$now = DateTime::createFromFormat('U.u', number_format(microtime(true), 6, '.', ''));
+					$nowFormatted = str_replace(["T", "Z"], " ", trim(substr($now->format('Y-m-d\TH:i:s.u'), 0, -3) . 'Z'));
 					$reviewUpdateResult = $link->query("UPDATE reviews SET review_text='$reviewText', rating='$rating', is_anonymous='$isAnonymous', create_datetime='$nowFormatted' 
                									WHERE user_id='$userID' AND movie_id='$movieID' AND review_id='$reviewID'");
 				}
@@ -115,25 +86,16 @@
 				}
 				$token = substr(getallheaders()['Authorization'], 7);
 				$isLogoutToken = $link->query("SELECT user_id FROM tokens WHERE value LIKE '$token'")->fetch_assoc();
-				if (!isExpired($token) && $isLogoutToken == null) {
+				if (isValid($token) && !isExpired($token) && $isLogoutToken == null) {
 					$usernameFromToken = getPayload($token)['unique_name'];
 					$user = $link->query("SELECT user_id FROM users WHERE username='$usernameFromToken'")->fetch_assoc();
 					$userID = $user['user_id'];
 					$movieID = $urlList[2];
 					$reviewID = $urlList[4];
-					$reviewCheck = $link->query("SELECT review_id FROM reviews WHERE review_id='$reviewID' AND user_id='$userID'")->fetch_assoc();
-					$movieCheck = $link->query("SELECT review_id FROM reviews WHERE user_id='$userID' AND movie_id='$movieID'")->fetch_assoc();
-					if (!$movieCheck) {
-						setHTTPStatus('404','The user has no review on this movie');
+					if (!checkIfTheUserHasReview($movieID, $userID) || !checkIfTheUserCanChangeReview($reviewID, $userID)) {
 						return;
 					}
-					if ($reviewCheck) {
-						$reviewDeleteResult = $link->query("DELETE FROM reviews WHERE user_id='$userID' AND movie_id='$movieID' AND review_id='$reviewID'");
-					}
-					else {
-						setHTTPStatus('403','User cannot delete foreign review');
-						return;
-					}
+					$reviewDeleteResult = $link->query("DELETE FROM reviews WHERE user_id='$userID' AND movie_id='$movieID' AND review_id='$reviewID'");
 				}
 				else {
 					setHTTPStatus('401', 'Token not specified or not valid');
